+++ kubejs/server_scripts/recipes/enigmatica/remove.js
@@ -16,7 +16,10 @@ ServerEvents.recipes((event) => {
         { id: /powah:crafting\/cable_(basic|hardened|blazing|niotic|spirited|nitro)$/ },
         { mod: 'cobblegengalore' },
         { id: /actuallyadditions:crushing\/(copper|iron|gold)_ore/ },
-        { id: /industrialforegoing:laser_drill_ore.*yellor(ite|ium)/ }
+        { id: /industrialforegoing:laser_drill_ore.*yellor(ite|ium)/ },
+        { id: `utilitarian:utility/green_dye` },
+        { id: `enderio:tank_fill/experience_bottle` },
+        { id: `industrialforegoing:dissolution_chamber/xp_bottles` }
     ];
 
     recipes.forEach((recipe) => {